requirements.txt
amqp==1.4.8
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.22
celery==3.1.19
Django==1.8.6
django-appconf==1.0.1
django-celery==3.1.17
django-ckeditor==5.0.3
django-filter==0.11.0
django-htmlmin==0.9.0
djangorestframework==3.3.1
gunicorn==19.4.1
html5lib==0.9999999
kombu==3.0.32
Markdown==2.6.5
mysqlclient==1.3.7
pbr==1.8.1
Pillow==3.0.0
python-memcached==1.57
pytz==2015.7
requests==2.8.1
six==1.10.0
stevedore==1.10.0
virtualenv==14.0.5
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.24.0
